Move exports to exports.ts
