+++ src/index.ts
@@ -18,14 +18,6 @@ import * as fs from 'fs';
 import * as SchemaDefs from './schema-defs/MatcherConf';
 import { startMatcher } from './exports';
 import { logger } from './Logger';
-import * as MatcherLogic from './matcher/MatcherLogic';
-
-const METHOD_NOT_IMPLEMENTED = 'Method not implemented.';
-
-export {
-    MatcherLogic,
-    METHOD_NOT_IMPLEMENTED
-};
 
 const main = async () => {
     if (process.argv[2]) {
